+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ SFA
    * Custom UI error pages
    * Deployment error codes explanation
    * "reboot_controller" special deployment API function
+   * serving public and private structured data from EVA ICS Registry
 
 
 3.3.2 (2021-03-01)